Rename default collection
The PBSHM Core can have multiple structure collections within a single
database. This facilitates the premise of data silos, where each realm
of confidential data has its own silo (a structure collection) for structure
data to reside within. Currently, the name 'structure_collection' is
misleading, as there may be multiple structure collections the user
has access to. As such, it is renamed to default_collection to denote
that it is only the default data silos and there may be more available.