Fix loading user schema
Using open_resources to load a file, uses a relative path from
the flask app root directory; however, when the application is deployed
through a site package, the location of the pbshm namespace
may not be in the flask app directory. As such, the file needs to be
loaded via the standard open method by determining the
directory of the initialisation module and then appending the
relative path of the file.